fix warden tests
